requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
barcode==1.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.33.0
Django==4.2.2
django-celery-results==2.5.1
django-db==0.0.7
django-mathfilters==1.0.0
django-onesignal==1.0.0
django-otp==1.2.2
django-sesame==3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
kombu==5.3.1
Levenshtein==0.21.1
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.2.0
Pillow==9.5.0
plyer==2.1.0
prompt-toolkit==3.0.38
protobuf==3.20.3

pyasn1==0.5.0
pycrypto==2.6.1
PyJWT==2.7.0
PyMySQL==1.1.0
pypng==0.20220715.0

python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-keyczar==0.716
python-Levenshtein==0.21.1
python-magic==0.4.27
pytz==2023.3
pywebio==1.8.2
PyYAML==6.0

rapidfuzz==3.1.1
requests==2.31.0
sesame==0.3.3
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2

typing_extensions==4.6.3
tzdata==2023.3
ua-parser==0.18.0
urllib3==1.26.16
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
yarl==1.9.2
